Move api/server/router/build to daemon/server/router/build
