+++ api/server/router/system/system_routes.go
@@ -12,7 +12,6 @@ import (
 
 	"github.com/containerd/log"
 	"github.com/docker/docker/api/server/httputils"
-	"github.com/docker/docker/api/server/router/build"
 	"github.com/docker/docker/api/types"
 	buildtypes "github.com/docker/docker/api/types/build"
 	"github.com/docker/docker/api/types/events"
@@ -22,6 +21,7 @@ import (
 	"github.com/docker/docker/api/types/system"
 	timetypes "github.com/docker/docker/api/types/time"
 	"github.com/docker/docker/api/types/versions"
+	"github.com/docker/docker/daemon/server/router/build"
 	"github.com/docker/docker/pkg/ioutils"
 	"github.com/pkg/errors"
 	"golang.org/x/sync/errgroup"
