+++ daemon/cluster/executor/container/validate_windows_test.go
@@ -1,6 +1,7 @@
 //go:build windows
 
 package container
+
 import (
 	"strings"
 	"testing"
