+++ daemon/command/daemon.go
@@ -23,7 +23,6 @@ import (
 	buildbackend "github.com/docker/docker/api/server/backend/build"
 	"github.com/docker/docker/api/server/middleware"
 	"github.com/docker/docker/api/server/router"
-	"github.com/docker/docker/api/server/router/build"
 	checkpointrouter "github.com/docker/docker/api/server/router/checkpoint"
 	"github.com/docker/docker/api/server/router/container"
 	debugrouter "github.com/docker/docker/api/server/router/debug"
@@ -45,6 +44,7 @@ import (
 	"github.com/docker/docker/daemon/command/trap"
 	"github.com/docker/docker/daemon/config"
 	"github.com/docker/docker/daemon/listeners"
+	"github.com/docker/docker/daemon/server/router/build"
 	"github.com/docker/docker/dockerversion"
 	"github.com/docker/docker/internal/otelutil"
 	"github.com/docker/docker/libcontainerd/supervisor"
